@@ -4,5 +4,5 @@ var searchData=
   ['jsonarraywrapper_1',['JSONArrayWrapper',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html',1,'json::utility']]],
   ['jsonbuilder_2',['jsonbuilder',['../classjson_1_1_j_s_o_n_builder.html',1,'json::JSONBuilder'],['../classjson_1_1_j_s_o_n_builder.html#a458f3dd94d29534c9f8315e4ecd445bd',1,'json::JSONBuilder::JSONBuilder(uint32_t codePage, outputType type=outputType::standard)'],['../classjson_1_1_j_s_o_n_builder.html#a787ed68408d33b05866b7d2686c06d4e',1,'json::JSONBuilder::JSONBuilder(const utility::jsonObject &amp;data, uint32_t codePage, outputType type=outputType::standard)'],['../classjson_1_1_j_s_o_n_builder.html#a427a84220bb6839a6842fe58715f0c68',1,'json::JSONBuilder::JSONBuilder(const JSONBuilder &amp;other)'],['../classjson_1_1_j_s_o_n_builder.html#a1905e2af447eff01fb14791565006789',1,'json::JSONBuilder::JSONBuilder(JSONBuilder &amp;&amp;other) noexcept']]],
   ['jsonobject_3',['jsonobject',['../classjson_1_1utility_1_1json_object.html',1,'json::utility::jsonObject'],['../classjson_1_1utility_1_1json_object.html#ad380414c04269e67093770a9ac55f78c',1,'json::utility::jsonObject::jsonObject(const jsonObject &amp;other)'],['../classjson_1_1utility_1_1json_object.html#a9482b3512c268e6ba157c351f6829b71',1,'json::utility::jsonObject::jsonObject(jsonObject &amp;&amp;other) noexcept']]],
-  ['jsonparser_4',['jsonparser',['../classjson_1_1_j_s_o_n_parser.html',1,'json::JSONParser'],['../classjson_1_1_j_s_o_n_parser.html#af850c82aa16b8620ea514209a5374bb2',1,'json::JSONParser::JSONParser(const std::string &amp;data)'],['../classjson_1_1_j_s_o_n_parser.html#aebacadf170056bcd9df95310de946d43',1,'json::JSONParser::JSONParser(std::istream &amp;inputStream)'],['../classjson_1_1_j_s_o_n_parser.html#a2b169c81ad50600f5a1fb2708d373ab3',1,'json::JSONParser::JSONParser(std::istream &amp;&amp;inputStream)'],['../classjson_1_1_j_s_o_n_parser.html#aa3e5cd1f5d707c21ff0ddf8e11f64190',1,'json::JSONParser::JSONParser(const utility::jsonObject &amp;data)'],['../classjson_1_1_j_s_o_n_parser.html#a12e9da81defc49b433bc104f10f8464c',1,'json::JSONParser::JSONParser(const JSONParser &amp;other)'],['../classjson_1_1_j_s_o_n_parser.html#a38cc5a98bd3d9ab79dd1d6d14c16dad5',1,'json::JSONParser::JSONParser(JSONParser &amp;&amp;other) noexcept']]]
+  ['jsonparser_4',['jsonparser',['../classjson_1_1_j_s_o_n_parser.html',1,'json::JSONParser'],['../classjson_1_1_j_s_o_n_parser.html#a1bc4e344187898d70e64ec6db0185eef',1,'json::JSONParser::JSONParser(std::string_view data)'],['../classjson_1_1_j_s_o_n_parser.html#aebacadf170056bcd9df95310de946d43',1,'json::JSONParser::JSONParser(std::istream &amp;inputStream)'],['../classjson_1_1_j_s_o_n_parser.html#a2b169c81ad50600f5a1fb2708d373ab3',1,'json::JSONParser::JSONParser(std::istream &amp;&amp;inputStream)'],['../classjson_1_1_j_s_o_n_parser.html#aa3e5cd1f5d707c21ff0ddf8e11f64190',1,'json::JSONParser::JSONParser(const utility::jsonObject &amp;data)'],['../classjson_1_1_j_s_o_n_parser.html#a12e9da81defc49b433bc104f10f8464c',1,'json::JSONParser::JSONParser(const JSONParser &amp;other)'],['../classjson_1_1_j_s_o_n_parser.html#a38cc5a98bd3d9ab79dd1d6d14c16dad5',1,'json::JSONParser::JSONParser(JSONParser &amp;&amp;other) noexcept']]]
 ];
