@@ -1,13 +1,13 @@
 var searchData=
 [
-  ['setarray_0',['setarray',['../classjson_1_1utility_1_1json_object.html#afd52eafb47e65dee4313d6ebb429de77',1,'json::utility::jsonObject::setArray(const std::string &amp;key, const std::vector&lt; jsonObject &gt; &amp;value)'],['../classjson_1_1utility_1_1json_object.html#a7bec7a16a0fa4cf7fbe7ba8efeb342ab',1,'json::utility::jsonObject::setArray(std::string &amp;&amp;key, const std::vector&lt; jsonObject &gt; &amp;value)'],['../classjson_1_1utility_1_1json_object.html#a3d6cbb7fd405a0b4133e4199a42c42cf',1,'json::utility::jsonObject::setArray(const std::string &amp;key, std::vector&lt; jsonObject &gt; &amp;&amp;value)'],['../classjson_1_1utility_1_1json_object.html#a9427b0184b7dc8ab1527e99422fe44fa',1,'json::utility::jsonObject::setArray(std::string &amp;&amp;key, std::vector&lt; jsonObject &gt; &amp;&amp;value)']]],
-  ['setbool_1',['setbool',['../classjson_1_1utility_1_1json_object.html#a28433e063d51d089a8b546e167b19222',1,'json::utility::jsonObject::setBool(const std::string &amp;key, bool value)'],['../classjson_1_1utility_1_1json_object.html#ad75e68ccf8d8b9ab1fd425f851348ac5',1,'json::utility::jsonObject::setBool(std::string &amp;&amp;key, bool value)']]],
-  ['setdouble_2',['setdouble',['../classjson_1_1utility_1_1json_object.html#adcfd4efcc44f07a771ae7c59b48f0d4b',1,'json::utility::jsonObject::setDouble(const std::string &amp;key, double value)'],['../classjson_1_1utility_1_1json_object.html#a02b353d126849273c0e667a4960d99fe',1,'json::utility::jsonObject::setDouble(std::string &amp;&amp;key, double value)']]],
-  ['setint_3',['setint',['../classjson_1_1utility_1_1json_object.html#a25f74b2d4b6d3d3a3f9401791697c31f',1,'json::utility::jsonObject::setInt(const std::string &amp;key, int64_t value)'],['../classjson_1_1utility_1_1json_object.html#aa8489ed73d285ba0d172578706e3f460',1,'json::utility::jsonObject::setInt(std::string &amp;&amp;key, int64_t value)']]],
-  ['setjsondata_4',['setjsondata',['../classjson_1_1_j_s_o_n_parser.html#a19fe9231149199232c9b81986ec5b983',1,'json::JSONParser::setJSONData(const std::string &amp;jsonData, uint32_t codePage)'],['../classjson_1_1_j_s_o_n_parser.html#a120bec0db8f7e423ac8f700c3b33f5a2',1,'json::JSONParser::setJSONData(const std::string &amp;jsonData)'],['../classjson_1_1_j_s_o_n_parser.html#a4878fe23b655bdc6b87296875bb4ec88',1,'json::JSONParser::setJSONData(std::string &amp;&amp;jsonData) noexcept'],['../classjson_1_1_j_s_o_n_parser.html#a29e800cbdef3b530f108be8aa4450efe',1,'json::JSONParser::setJSONData(std::istream &amp;inputStream)'],['../classjson_1_1_j_s_o_n_parser.html#abad51b30cac1d5948dce5fa761d45c75',1,'json::JSONParser::setJSONData(std::istream &amp;&amp;inputStream)']]],
-  ['setnull_5',['setnull',['../classjson_1_1utility_1_1json_object.html#a38fd72d647ec4338ec4639d7d7b91268',1,'json::utility::jsonObject::setNull(const std::string &amp;key)'],['../classjson_1_1utility_1_1json_object.html#afea5d32506bced68aaaf04c10b25fe10',1,'json::utility::jsonObject::setNull(std::string &amp;&amp;key)']]],
-  ['setobject_6',['setobject',['../classjson_1_1utility_1_1json_object.html#a4ec55140ac6f96ddec48b0ad1c80e0a0',1,'json::utility::jsonObject::setObject(const std::string &amp;key, const jsonObject &amp;value)'],['../classjson_1_1utility_1_1json_object.html#ab53545e9f06f1816b2957dfe70b9f11a',1,'json::utility::jsonObject::setObject(std::string &amp;&amp;key, const jsonObject &amp;value)'],['../classjson_1_1utility_1_1json_object.html#af12e87025aa2935b89ec8bf8b4e0afa7',1,'json::utility::jsonObject::setObject(const std::string &amp;key, jsonObject &amp;&amp;value)'],['../classjson_1_1utility_1_1json_object.html#a892221be33efd2082d18b3c5360c7bec',1,'json::utility::jsonObject::setObject(std::string &amp;&amp;key, jsonObject &amp;&amp;value)']]],
-  ['setstring_7',['setstring',['../classjson_1_1utility_1_1json_object.html#a3eb9f5b6f47bde0ed6668a0c12d63201',1,'json::utility::jsonObject::setString(const std::string &amp;key, const std::string &amp;value)'],['../classjson_1_1utility_1_1json_object.html#a2a72519e54f435611757dba148a0e746',1,'json::utility::jsonObject::setString(std::string &amp;&amp;key, const std::string &amp;value)'],['../classjson_1_1utility_1_1json_object.html#abfed87b0a6a510467ac50393ce53b0c0',1,'json::utility::jsonObject::setString(const std::string &amp;key, std::string &amp;&amp;value)'],['../classjson_1_1utility_1_1json_object.html#a2dd2d67e5c965e44e8a45e12e62a47ef',1,'json::utility::jsonObject::setString(std::string &amp;&amp;key, std::string &amp;&amp;value)']]],
-  ['setunsignedint_8',['setunsignedint',['../classjson_1_1utility_1_1json_object.html#add836f1ab5bc7397243df440dba65386',1,'json::utility::jsonObject::setUnsignedInt(const std::string &amp;key, uint64_t value)'],['../classjson_1_1utility_1_1json_object.html#a1838ac38ce2449b8db6308968e2e1d59',1,'json::utility::jsonObject::setUnsignedInt(std::string &amp;&amp;key, uint64_t value)']]],
+  ['setarray_0',['setarray',['../classjson_1_1utility_1_1json_object.html#a36143fdc97fd8363f7b478246635a8ed',1,'json::utility::jsonObject::setArray(std::string_view key, const std::vector&lt; jsonObject &gt; &amp;value)'],['../classjson_1_1utility_1_1json_object.html#a7bec7a16a0fa4cf7fbe7ba8efeb342ab',1,'json::utility::jsonObject::setArray(std::string &amp;&amp;key, const std::vector&lt; jsonObject &gt; &amp;value)'],['../classjson_1_1utility_1_1json_object.html#a3c096642b2d573c4d468baef35e51690',1,'json::utility::jsonObject::setArray(std::string_view key, std::vector&lt; jsonObject &gt; &amp;&amp;value)'],['../classjson_1_1utility_1_1json_object.html#a9427b0184b7dc8ab1527e99422fe44fa',1,'json::utility::jsonObject::setArray(std::string &amp;&amp;key, std::vector&lt; jsonObject &gt; &amp;&amp;value)']]],
+  ['setbool_1',['setbool',['../classjson_1_1utility_1_1json_object.html#a8f45d03f5bede54ca785f21fdc11ae41',1,'json::utility::jsonObject::setBool(std::string_view key, bool value)'],['../classjson_1_1utility_1_1json_object.html#ad75e68ccf8d8b9ab1fd425f851348ac5',1,'json::utility::jsonObject::setBool(std::string &amp;&amp;key, bool value)']]],
+  ['setdouble_2',['setdouble',['../classjson_1_1utility_1_1json_object.html#abf8cc0b2b4cfd44b4838ee7518e160bf',1,'json::utility::jsonObject::setDouble(std::string_view key, double value)'],['../classjson_1_1utility_1_1json_object.html#a02b353d126849273c0e667a4960d99fe',1,'json::utility::jsonObject::setDouble(std::string &amp;&amp;key, double value)']]],
+  ['setint_3',['setint',['../classjson_1_1utility_1_1json_object.html#a4a9b9c36e0cdb409168ba47a8cb1655d',1,'json::utility::jsonObject::setInt(std::string_view key, int64_t value)'],['../classjson_1_1utility_1_1json_object.html#aa8489ed73d285ba0d172578706e3f460',1,'json::utility::jsonObject::setInt(std::string &amp;&amp;key, int64_t value)']]],
+  ['setjsondata_4',['setjsondata',['../classjson_1_1_j_s_o_n_parser.html#add8b35c03524a19de42e3ad9fb97d04c',1,'json::JSONParser::setJSONData(std::string_view jsonData, uint32_t codePage)'],['../classjson_1_1_j_s_o_n_parser.html#a82abad11700168b4bb994174fcabe7b8',1,'json::JSONParser::setJSONData(std::string_view jsonData)'],['../classjson_1_1_j_s_o_n_parser.html#a4878fe23b655bdc6b87296875bb4ec88',1,'json::JSONParser::setJSONData(std::string &amp;&amp;jsonData) noexcept'],['../classjson_1_1_j_s_o_n_parser.html#a29e800cbdef3b530f108be8aa4450efe',1,'json::JSONParser::setJSONData(std::istream &amp;inputStream)'],['../classjson_1_1_j_s_o_n_parser.html#abad51b30cac1d5948dce5fa761d45c75',1,'json::JSONParser::setJSONData(std::istream &amp;&amp;inputStream)']]],
+  ['setnull_5',['setnull',['../classjson_1_1utility_1_1json_object.html#a805b815ee32ffadd703afe20827d6b6e',1,'json::utility::jsonObject::setNull(std::string_view key)'],['../classjson_1_1utility_1_1json_object.html#afea5d32506bced68aaaf04c10b25fe10',1,'json::utility::jsonObject::setNull(std::string &amp;&amp;key)']]],
+  ['setobject_6',['setobject',['../classjson_1_1utility_1_1json_object.html#aaa17b26aed63c10cb287760559ce1ae6',1,'json::utility::jsonObject::setObject(std::string_view key, const jsonObject &amp;value)'],['../classjson_1_1utility_1_1json_object.html#ab53545e9f06f1816b2957dfe70b9f11a',1,'json::utility::jsonObject::setObject(std::string &amp;&amp;key, const jsonObject &amp;value)'],['../classjson_1_1utility_1_1json_object.html#aba52e50c5572d26e7308ec2e32bcc32b',1,'json::utility::jsonObject::setObject(std::string_view key, jsonObject &amp;&amp;value)'],['../classjson_1_1utility_1_1json_object.html#a892221be33efd2082d18b3c5360c7bec',1,'json::utility::jsonObject::setObject(std::string &amp;&amp;key, jsonObject &amp;&amp;value)']]],
+  ['setstring_7',['setstring',['../classjson_1_1utility_1_1json_object.html#a94b6eb7a66bbe05e21dbb9403586f7e9',1,'json::utility::jsonObject::setString(std::string_view key, std::string_view value)'],['../classjson_1_1utility_1_1json_object.html#a6611f88825d11298df0683cf86e9350b',1,'json::utility::jsonObject::setString(std::string &amp;&amp;key, std::string_view value)'],['../classjson_1_1utility_1_1json_object.html#a66a2aed57339a377a583859501321f22',1,'json::utility::jsonObject::setString(std::string_view key, std::string &amp;&amp;value)'],['../classjson_1_1utility_1_1json_object.html#a2dd2d67e5c965e44e8a45e12e62a47ef',1,'json::utility::jsonObject::setString(std::string &amp;&amp;key, std::string &amp;&amp;value)']]],
+  ['setunsignedint_8',['setunsignedint',['../classjson_1_1utility_1_1json_object.html#ae6a5869cbaffc6c3d97c2bf70a579fe4',1,'json::utility::jsonObject::setUnsignedInt(std::string_view key, uint64_t value)'],['../classjson_1_1utility_1_1json_object.html#a1838ac38ce2449b8db6308968e2e1d59',1,'json::utility::jsonObject::setUnsignedInt(std::string &amp;&amp;key, uint64_t value)']]],
   ['standard_9',['standard',['../classjson_1_1_j_s_o_n_builder.html#abd4f82f0a0bc9008b89c634faf810573',1,'json::JSONBuilder::standard()'],['../classjson_1_1_j_s_o_n_builder.html#a613fcb5202d9a8512787fa10040ef6b8ac00f0c4675b91fb8b918e4079a0b1bac',1,'json::JSONBuilder::standard']]]
 ];
