@@ -4,40 +4,48 @@ var classjson_1_1utility_1_1json_object =
     [ "jsonObject", "classjson_1_1utility_1_1json_object.html#ad380414c04269e67093770a9ac55f78c", null ],
     [ "jsonObject", "classjson_1_1utility_1_1json_object.html#a9482b3512c268e6ba157c351f6829b71", null ],
     [ "begin", "classjson_1_1utility_1_1json_object.html#ae60c1e1f744c71dd9311df6fe9cf9873", null ],
-    [ "contains", "classjson_1_1utility_1_1json_object.html#a82052adba91df4500a4cb41c43d18132", null ],
+    [ "contains", "classjson_1_1utility_1_1json_object.html#a2aedee9c3047907f882218dbbd2ffafc", null ],
     [ "end", "classjson_1_1utility_1_1json_object.html#a3628f29bf4d92a69400a2d60df176f73", null ],
-    [ "getArray", "classjson_1_1utility_1_1json_object.html#ab1486715ff975a002f9d2041bd35f427", null ],
-    [ "getBool", "classjson_1_1utility_1_1json_object.html#a27b161e7c92fc41b2053281b275b8adb", null ],
-    [ "getDouble", "classjson_1_1utility_1_1json_object.html#a8d55483e4b2113652d4318d6a210910d", null ],
-    [ "getInt", "classjson_1_1utility_1_1json_object.html#a5b0954200ef06a70e2cf9b3a4afdf3db", null ],
-    [ "getNull", "classjson_1_1utility_1_1json_object.html#a5fa34da0933d80241946a7eb6d361d1c", null ],
-    [ "getObject", "classjson_1_1utility_1_1json_object.html#a598e2050cbadc5c695c7737302193b51", null ],
-    [ "getString", "classjson_1_1utility_1_1json_object.html#a60c5e1e45ea291be5f3aa67d143d61ce", null ],
-    [ "getUnsignedInt", "classjson_1_1utility_1_1json_object.html#ac4c9d994ce1347043481306e1b164bdb", null ],
+    [ "getArray", "classjson_1_1utility_1_1json_object.html#a376b70a297c08788a38b54bd9c64ee8e", null ],
+    [ "getBool", "classjson_1_1utility_1_1json_object.html#a61d5124942814d04b15c2559662cb299", null ],
+    [ "getDouble", "classjson_1_1utility_1_1json_object.html#a199c972a61a209dffdff84d0438671ad", null ],
+    [ "getInt", "classjson_1_1utility_1_1json_object.html#a45b188c694310cc505a979b635d1f6f2", null ],
+    [ "getNull", "classjson_1_1utility_1_1json_object.html#a9cdd39a00489790321c857b7b5492a10", null ],
+    [ "getObject", "classjson_1_1utility_1_1json_object.html#a49d5388f6c2b1f7add9df2118ce955ce", null ],
+    [ "getString", "classjson_1_1utility_1_1json_object.html#ac92e7288687ceefdfe0a25aa496f70d1", null ],
+    [ "getUnsignedInt", "classjson_1_1utility_1_1json_object.html#a959fe35ffc1f861f1b84885bab8848a6", null ],
     [ "operator=", "classjson_1_1utility_1_1json_object.html#a6ed3e97aa6075b874f9044e60fd7debe", null ],
     [ "operator=", "classjson_1_1utility_1_1json_object.html#a433f9712f3ba4c0a0444c61b6133574b", null ],
-    [ "operator[]", "classjson_1_1utility_1_1json_object.html#adfc8fe04ecef682d882aee70ec146bfd", null ],
-    [ "operator[]", "classjson_1_1utility_1_1json_object.html#a0ced305bd837c4aae362fb2f3fa604c3", null ],
-    [ "setArray", "classjson_1_1utility_1_1json_object.html#afd52eafb47e65dee4313d6ebb429de77", null ],
-    [ "setArray", "classjson_1_1utility_1_1json_object.html#a3d6cbb7fd405a0b4133e4199a42c42cf", null ],
+    [ "operator[]", "classjson_1_1utility_1_1json_object.html#a5bba7f7ec1a0680ba43cf177fadc11c7", null ],
+    [ "operator[]", "classjson_1_1utility_1_1json_object.html#a91454b0ccbf9e3016bb9d1a3a7f3e409", null ],
     [ "setArray", "classjson_1_1utility_1_1json_object.html#a7bec7a16a0fa4cf7fbe7ba8efeb342ab", null ],
     [ "setArray", "classjson_1_1utility_1_1json_object.html#a9427b0184b7dc8ab1527e99422fe44fa", null ],
-    [ "setBool", "classjson_1_1utility_1_1json_object.html#a28433e063d51d089a8b546e167b19222", null ],
+    [ "setArray", "classjson_1_1utility_1_1json_object.html#a36143fdc97fd8363f7b478246635a8ed", null ],
+    [ "setArray", "classjson_1_1utility_1_1json_object.html#a3c096642b2d573c4d468baef35e51690", null ],
     [ "setBool", "classjson_1_1utility_1_1json_object.html#ad75e68ccf8d8b9ab1fd425f851348ac5", null ],
-    [ "setDouble", "classjson_1_1utility_1_1json_object.html#adcfd4efcc44f07a771ae7c59b48f0d4b", null ],
+    [ "setBool", "classjson_1_1utility_1_1json_object.html#a8f45d03f5bede54ca785f21fdc11ae41", null ],
     [ "setDouble", "classjson_1_1utility_1_1json_object.html#a02b353d126849273c0e667a4960d99fe", null ],
-    [ "setInt", "classjson_1_1utility_1_1json_object.html#a25f74b2d4b6d3d3a3f9401791697c31f", null ],
+    [ "setDouble", "classjson_1_1utility_1_1json_object.html#abf8cc0b2b4cfd44b4838ee7518e160bf", null ],
     [ "setInt", "classjson_1_1utility_1_1json_object.html#aa8489ed73d285ba0d172578706e3f460", null ],
-    [ "setNull", "classjson_1_1utility_1_1json_object.html#a38fd72d647ec4338ec4639d7d7b91268", null ],
+    [ "setInt", "classjson_1_1utility_1_1json_object.html#a4a9b9c36e0cdb409168ba47a8cb1655d", null ],
     [ "setNull", "classjson_1_1utility_1_1json_object.html#afea5d32506bced68aaaf04c10b25fe10", null ],
-    [ "setObject", "classjson_1_1utility_1_1json_object.html#a4ec55140ac6f96ddec48b0ad1c80e0a0", null ],
-    [ "setObject", "classjson_1_1utility_1_1json_object.html#af12e87025aa2935b89ec8bf8b4e0afa7", null ],
+    [ "setNull", "classjson_1_1utility_1_1json_object.html#a805b815ee32ffadd703afe20827d6b6e", null ],
     [ "setObject", "classjson_1_1utility_1_1json_object.html#ab53545e9f06f1816b2957dfe70b9f11a", null ],
     [ "setObject", "classjson_1_1utility_1_1json_object.html#a892221be33efd2082d18b3c5360c7bec", null ],
-    [ "setString", "classjson_1_1utility_1_1json_object.html#a3eb9f5b6f47bde0ed6668a0c12d63201", null ],
-    [ "setString", "classjson_1_1utility_1_1json_object.html#abfed87b0a6a510467ac50393ce53b0c0", null ],
-    [ "setString", "classjson_1_1utility_1_1json_object.html#a2a72519e54f435611757dba148a0e746", null ],
+    [ "setObject", "classjson_1_1utility_1_1json_object.html#aaa17b26aed63c10cb287760559ce1ae6", null ],
+    [ "setObject", "classjson_1_1utility_1_1json_object.html#aba52e50c5572d26e7308ec2e32bcc32b", null ],
     [ "setString", "classjson_1_1utility_1_1json_object.html#a2dd2d67e5c965e44e8a45e12e62a47ef", null ],
-    [ "setUnsignedInt", "classjson_1_1utility_1_1json_object.html#add836f1ab5bc7397243df440dba65386", null ],
-    [ "setUnsignedInt", "classjson_1_1utility_1_1json_object.html#a1838ac38ce2449b8db6308968e2e1d59", null ]
+    [ "setString", "classjson_1_1utility_1_1json_object.html#a6611f88825d11298df0683cf86e9350b", null ],
+    [ "setString", "classjson_1_1utility_1_1json_object.html#a66a2aed57339a377a583859501321f22", null ],
+    [ "setString", "classjson_1_1utility_1_1json_object.html#a94b6eb7a66bbe05e21dbb9403586f7e9", null ],
+    [ "setUnsignedInt", "classjson_1_1utility_1_1json_object.html#a1838ac38ce2449b8db6308968e2e1d59", null ],
+    [ "setUnsignedInt", "classjson_1_1utility_1_1json_object.html#ae6a5869cbaffc6c3d97c2bf70a579fe4", null ],
+    [ "tryGetArray", "classjson_1_1utility_1_1json_object.html#aaaa43ff0df2272147af9173f642d7330", null ],
+    [ "tryGetBool", "classjson_1_1utility_1_1json_object.html#a338a864c3c7af2238a8fbe8c3b5f5408", null ],
+    [ "tryGetDouble", "classjson_1_1utility_1_1json_object.html#a123579a23dc8adad5bdaa477725fd090", null ],
+    [ "tryGetInt", "classjson_1_1utility_1_1json_object.html#a5eebc4e04cba6f7d250cf6b4009d6b72", null ],
+    [ "tryGetNull", "classjson_1_1utility_1_1json_object.html#a28579c08743212a69a5cb2c5f4ee3a92", null ],
+    [ "tryGetObject", "classjson_1_1utility_1_1json_object.html#a41ace1fa6383acd60e0544eebe136533", null ],
+    [ "tryGetString", "classjson_1_1utility_1_1json_object.html#abd2dd20c04a47a19ad0e5e42857a4a1d", null ],
+    [ "tryGetUnsignedInt", "classjson_1_1utility_1_1json_object.html#a76a6b07b18dde76fdefe0f445bab5582", null ]
 ];
