@@ -52,46 +52,46 @@ var NAVTREEINDEX0 =
 "classjson_1_1_j_s_o_n_builder.html#af9a4550b4e977436ce5e93be0dcbd628":[1,0,0,2,27],
 "classjson_1_1_j_s_o_n_parser.html":[1,0,0,3],
 "classjson_1_1_j_s_o_n_parser.html#a01395b13e66011b067567d97bedea32c":[1,0,0,3,17],
-"classjson_1_1_j_s_o_n_parser.html#a09c38da50018f7e9cd5310d05630a05e":[1,0,0,3,12],
-"classjson_1_1_j_s_o_n_parser.html#a0d5793e89a23798caf84ea484dbbaaa7":[1,0,0,3,11],
-"classjson_1_1_j_s_o_n_parser.html#a120bec0db8f7e423ac8f700c3b33f5a2":[1,0,0,3,25],
 "classjson_1_1_j_s_o_n_parser.html#a12e9da81defc49b433bc104f10f8464c":[1,0,0,3,7],
-"classjson_1_1_j_s_o_n_parser.html#a188fd21d6c6769e1465a15186cf853d4":[1,0,0,3,31],
-"classjson_1_1_j_s_o_n_parser.html#a19fe9231149199232c9b81986ec5b983":[1,0,0,3,26],
-"classjson_1_1_j_s_o_n_parser.html#a1a63aeb4c99a31409ad13e749fb63733":[1,0,0,3,34],
-"classjson_1_1_j_s_o_n_parser.html#a29e800cbdef3b530f108be8aa4450efe":[1,0,0,3,28],
+"classjson_1_1_j_s_o_n_parser.html#a1690f2adb5837e8724af24c11009db66":[1,0,0,3,10],
+"classjson_1_1_j_s_o_n_parser.html#a1bc4e344187898d70e64ec6db0185eef":[1,0,0,3,3],
+"classjson_1_1_j_s_o_n_parser.html#a27dabb852c5595ef61479630112579c4":[1,0,0,3,35],
+"classjson_1_1_j_s_o_n_parser.html#a29e800cbdef3b530f108be8aa4450efe":[1,0,0,3,26],
+"classjson_1_1_j_s_o_n_parser.html#a2a51366349730b9e7ea2da82603904d9":[1,0,0,3,21],
 "classjson_1_1_j_s_o_n_parser.html#a2b169c81ad50600f5a1fb2708d373ab3":[1,0,0,3,5],
+"classjson_1_1_j_s_o_n_parser.html#a2ebdf48b377a6dbe3f4764a6146c65b2":[1,0,0,3,13],
+"classjson_1_1_j_s_o_n_parser.html#a38b1754bcf80b96be4e7953003bffadf":[1,0,0,3,15],
 "classjson_1_1_j_s_o_n_parser.html#a38cc5a98bd3d9ab79dd1d6d14c16dad5":[1,0,0,3,8],
+"classjson_1_1_j_s_o_n_parser.html#a3a4ac7936573d278999e8b7bac514bb6":[1,0,0,3,33],
+"classjson_1_1_j_s_o_n_parser.html#a3e7447a67746f778a3587365c40e7332":[1,0,0,3,20],
+"classjson_1_1_j_s_o_n_parser.html#a40a72606768701b8d75cfe095f36848b":[1,0,0,3,31],
+"classjson_1_1_j_s_o_n_parser.html#a4442cb7ae9c2461523a035e560b65beb":[1,0,0,3,32],
 "classjson_1_1_j_s_o_n_parser.html#a4570d8aebfd60f9e02368ee5ba7feb03":[1,0,0,3,39],
-"classjson_1_1_j_s_o_n_parser.html#a45c820bff7925a98e53385cfc73b9e32":[1,0,0,3,33],
 "classjson_1_1_j_s_o_n_parser.html#a4638e39245740b5bd8168f5d8b428728":[1,0,0,3,16],
-"classjson_1_1_j_s_o_n_parser.html#a4878fe23b655bdc6b87296875bb4ec88":[1,0,0,3,29],
-"classjson_1_1_j_s_o_n_parser.html#a512fa66b3d1751e7b738351d19963870":[1,0,0,3,30],
+"classjson_1_1_j_s_o_n_parser.html#a4878fe23b655bdc6b87296875bb4ec88":[1,0,0,3,27],
+"classjson_1_1_j_s_o_n_parser.html#a52b34b90e0c81e4885791b792824e503":[1,0,0,3,12],
+"classjson_1_1_j_s_o_n_parser.html#a5a836eb7dfec5e8cef9c467643392120":[1,0,0,3,30],
+"classjson_1_1_j_s_o_n_parser.html#a670e7c83fe4af820d951277fd495d89f":[1,0,0,3,37],
 "classjson_1_1_j_s_o_n_parser.html#a686542fead8a129a1e5cc9c5f065f4ea":[1,0,0,3,22],
-"classjson_1_1_j_s_o_n_parser.html#a6cd09b58620722ad3b18e587e12d7d28":[1,0,0,3,20],
+"classjson_1_1_j_s_o_n_parser.html#a6dbac42e459305ba544643ff0ea6fa46":[1,0,0,3,34],
 "classjson_1_1_j_s_o_n_parser.html#a6e0f4eae74468bf94663938c50babd3c":[1,0,0,3,40],
-"classjson_1_1_j_s_o_n_parser.html#a744b8138a1cfd3ea53112dd316bf0959":[1,0,0,3,32],
-"classjson_1_1_j_s_o_n_parser.html#a846ebb23fe7b987cbc05da7af7cd58a0":[1,0,0,3,15],
+"classjson_1_1_j_s_o_n_parser.html#a75863b2f4e96354c6b3b9b24b8d44935":[1,0,0,3,36],
+"classjson_1_1_j_s_o_n_parser.html#a82abad11700168b4bb994174fcabe7b8":[1,0,0,3,28],
 "classjson_1_1_j_s_o_n_parser.html#a8cc621411c60e165716416f0909b3e88":[1,0,0,3,23],
-"classjson_1_1_j_s_o_n_parser.html#a8e5523ed249d032c259a12e27af6ce52":[1,0,0,3,9],
 "classjson_1_1_j_s_o_n_parser.html#a936a650fdaa8aa0931547f1d755e775a":[1,0,0,3,24],
-"classjson_1_1_j_s_o_n_parser.html#a96fbc3ab20618e40747e7c9ca41bf0f6":[1,0,0,3,21],
 "classjson_1_1_j_s_o_n_parser.html#a978acc40730a8d003ebcece6c513a662":[1,0,0,3,0],
 "classjson_1_1_j_s_o_n_parser.html#a98013c1d3c0661d58ba0fb4fca64f804":[1,0,0,3,2],
 "classjson_1_1_j_s_o_n_parser.html#aa3e5cd1f5d707c21ff0ddf8e11f64190":[1,0,0,3,6],
-"classjson_1_1_j_s_o_n_parser.html#aa8581ea1f34ce9ec778c21e077888c09":[1,0,0,3,37],
-"classjson_1_1_j_s_o_n_parser.html#ab1de4290badcc9c470511fe5e8f9056a":[1,0,0,3,35],
-"classjson_1_1_j_s_o_n_parser.html#ab9c08a5659362b930e18487ee2723045":[1,0,0,3,10],
-"classjson_1_1_j_s_o_n_parser.html#abad51b30cac1d5948dce5fa761d45c75":[1,0,0,3,27],
-"classjson_1_1_j_s_o_n_parser.html#abafb1bc5b07e36d067947a43b40272a0":[1,0,0,3,36],
-"classjson_1_1_j_s_o_n_parser.html#ac711fd5c6efdfaed2703de7f9b2125e3":[1,0,0,3,19],
+"classjson_1_1_j_s_o_n_parser.html#aa645009eaee379cbfddb19c31f82fb5a":[1,0,0,3,14],
+"classjson_1_1_j_s_o_n_parser.html#ab1853fe8165e23afae7d92728e9461d7":[1,0,0,3,9],
+"classjson_1_1_j_s_o_n_parser.html#ab1987fc0f8a9ab7cda6b72af3cde4a1c":[1,0,0,3,11],
+"classjson_1_1_j_s_o_n_parser.html#abad51b30cac1d5948dce5fa761d45c75":[1,0,0,3,25],
+"classjson_1_1_j_s_o_n_parser.html#acbbbbcfec6d3b5f1d60c8eff7250b902":[1,0,0,3,19],
 "classjson_1_1_j_s_o_n_parser.html#ad4795080e49dd316122d9b1d67772fb4":[1,0,0,3,1],
-"classjson_1_1_j_s_o_n_parser.html#adeb54d6f62f958bb1ce1247b65e6b2f6":[1,0,0,3,38],
+"classjson_1_1_j_s_o_n_parser.html#add8b35c03524a19de42e3ad9fb97d04c":[1,0,0,3,29],
 "classjson_1_1_j_s_o_n_parser.html#ae3382fde4943072afef448781533d241":[1,0,0,3,18],
 "classjson_1_1_j_s_o_n_parser.html#aebacadf170056bcd9df95310de946d43":[1,0,0,3,4],
-"classjson_1_1_j_s_o_n_parser.html#aebb676dc15e1fa1ba6977a3e4fa18159":[1,0,0,3,13],
-"classjson_1_1_j_s_o_n_parser.html#af320cb94a23053295e885e86edd6fa8a":[1,0,0,3,14],
-"classjson_1_1_j_s_o_n_parser.html#af850c82aa16b8620ea514209a5374bb2":[1,0,0,3,3],
+"classjson_1_1_j_s_o_n_parser.html#aee5fc5ea7cfec9c79d6a8186bb3f3ba1":[1,0,0,3,38],
 "classjson_1_1_recursive_j_s_o_n_iterator.html":[1,0,0,4],
 "classjson_1_1exceptions_1_1_base_j_s_o_n_exception.html":[1,0,0,0,0],
 "classjson_1_1exceptions_1_1_cant_find_value_exception.html":[1,0,0,0,1],
@@ -117,45 +117,53 @@ var NAVTREEINDEX0 =
 "classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#af1ebf097fce9476eac08ec7800b684c4":[1,0,0,1,0,17],
 "classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#afe23a16435f74e45eb26658b3b330fea":[1,0,0,1,0,13],
 "classjson_1_1utility_1_1json_object.html":[1,0,0,1,1],
-"classjson_1_1utility_1_1json_object.html#a02b353d126849273c0e667a4960d99fe":[1,0,0,1,1,25],
-"classjson_1_1utility_1_1json_object.html#a0ced305bd837c4aae362fb2f3fa604c3":[1,0,0,1,1,17],
-"classjson_1_1utility_1_1json_object.html#a1838ac38ce2449b8db6308968e2e1d59":[1,0,0,1,1,39],
-"classjson_1_1utility_1_1json_object.html#a25f74b2d4b6d3d3a3f9401791697c31f":[1,0,0,1,1,26],
-"classjson_1_1utility_1_1json_object.html#a27b161e7c92fc41b2053281b275b8adb":[1,0,0,1,1,7],
-"classjson_1_1utility_1_1json_object.html#a28433e063d51d089a8b546e167b19222":[1,0,0,1,1,22],
-"classjson_1_1utility_1_1json_object.html#a2a72519e54f435611757dba148a0e746":[1,0,0,1,1,36],
-"classjson_1_1utility_1_1json_object.html#a2dd2d67e5c965e44e8a45e12e62a47ef":[1,0,0,1,1,37],
+"classjson_1_1utility_1_1json_object.html#a02b353d126849273c0e667a4960d99fe":[1,0,0,1,1,24],
+"classjson_1_1utility_1_1json_object.html#a123579a23dc8adad5bdaa477725fd090":[1,0,0,1,1,42],
+"classjson_1_1utility_1_1json_object.html#a1838ac38ce2449b8db6308968e2e1d59":[1,0,0,1,1,38],
+"classjson_1_1utility_1_1json_object.html#a199c972a61a209dffdff84d0438671ad":[1,0,0,1,1,8],
+"classjson_1_1utility_1_1json_object.html#a28579c08743212a69a5cb2c5f4ee3a92":[1,0,0,1,1,44],
+"classjson_1_1utility_1_1json_object.html#a2aedee9c3047907f882218dbbd2ffafc":[1,0,0,1,1,4],
+"classjson_1_1utility_1_1json_object.html#a2dd2d67e5c965e44e8a45e12e62a47ef":[1,0,0,1,1,34],
+"classjson_1_1utility_1_1json_object.html#a338a864c3c7af2238a8fbe8c3b5f5408":[1,0,0,1,1,41],
+"classjson_1_1utility_1_1json_object.html#a36143fdc97fd8363f7b478246635a8ed":[1,0,0,1,1,20],
 "classjson_1_1utility_1_1json_object.html#a3628f29bf4d92a69400a2d60df176f73":[1,0,0,1,1,5],
-"classjson_1_1utility_1_1json_object.html#a38fd72d647ec4338ec4639d7d7b91268":[1,0,0,1,1,28],
-"classjson_1_1utility_1_1json_object.html#a3d6cbb7fd405a0b4133e4199a42c42cf":[1,0,0,1,1,19],
-"classjson_1_1utility_1_1json_object.html#a3eb9f5b6f47bde0ed6668a0c12d63201":[1,0,0,1,1,34],
+"classjson_1_1utility_1_1json_object.html#a376b70a297c08788a38b54bd9c64ee8e":[1,0,0,1,1,6],
+"classjson_1_1utility_1_1json_object.html#a3c096642b2d573c4d468baef35e51690":[1,0,0,1,1,21],
+"classjson_1_1utility_1_1json_object.html#a41ace1fa6383acd60e0544eebe136533":[1,0,0,1,1,45],
 "classjson_1_1utility_1_1json_object.html#a433f9712f3ba4c0a0444c61b6133574b":[1,0,0,1,1,15],
-"classjson_1_1utility_1_1json_object.html#a4ec55140ac6f96ddec48b0ad1c80e0a0":[1,0,0,1,1,30],
-"classjson_1_1utility_1_1json_object.html#a598e2050cbadc5c695c7737302193b51":[1,0,0,1,1,11],
-"classjson_1_1utility_1_1json_object.html#a5b0954200ef06a70e2cf9b3a4afdf3db":[1,0,0,1,1,9],
-"classjson_1_1utility_1_1json_object.html#a5fa34da0933d80241946a7eb6d361d1c":[1,0,0,1,1,10],
-"classjson_1_1utility_1_1json_object.html#a60c5e1e45ea291be5f3aa67d143d61ce":[1,0,0,1,1,12],
+"classjson_1_1utility_1_1json_object.html#a45b188c694310cc505a979b635d1f6f2":[1,0,0,1,1,9],
+"classjson_1_1utility_1_1json_object.html#a49d5388f6c2b1f7add9df2118ce955ce":[1,0,0,1,1,11],
+"classjson_1_1utility_1_1json_object.html#a4a9b9c36e0cdb409168ba47a8cb1655d":[1,0,0,1,1,27],
+"classjson_1_1utility_1_1json_object.html#a5bba7f7ec1a0680ba43cf177fadc11c7":[1,0,0,1,1,16],
+"classjson_1_1utility_1_1json_object.html#a5eebc4e04cba6f7d250cf6b4009d6b72":[1,0,0,1,1,43],
+"classjson_1_1utility_1_1json_object.html#a61d5124942814d04b15c2559662cb299":[1,0,0,1,1,7],
+"classjson_1_1utility_1_1json_object.html#a6611f88825d11298df0683cf86e9350b":[1,0,0,1,1,35],
+"classjson_1_1utility_1_1json_object.html#a66a2aed57339a377a583859501321f22":[1,0,0,1,1,36],
 "classjson_1_1utility_1_1json_object.html#a6ed3e97aa6075b874f9044e60fd7debe":[1,0,0,1,1,14],
-"classjson_1_1utility_1_1json_object.html#a7bec7a16a0fa4cf7fbe7ba8efeb342ab":[1,0,0,1,1,20],
-"classjson_1_1utility_1_1json_object.html#a82052adba91df4500a4cb41c43d18132":[1,0,0,1,1,4],
-"classjson_1_1utility_1_1json_object.html#a892221be33efd2082d18b3c5360c7bec":[1,0,0,1,1,33],
-"classjson_1_1utility_1_1json_object.html#a8d55483e4b2113652d4318d6a210910d":[1,0,0,1,1,8],
-"classjson_1_1utility_1_1json_object.html#a9427b0184b7dc8ab1527e99422fe44fa":[1,0,0,1,1,21],
+"classjson_1_1utility_1_1json_object.html#a76a6b07b18dde76fdefe0f445bab5582":[1,0,0,1,1,47],
+"classjson_1_1utility_1_1json_object.html#a7bec7a16a0fa4cf7fbe7ba8efeb342ab":[1,0,0,1,1,18],
+"classjson_1_1utility_1_1json_object.html#a805b815ee32ffadd703afe20827d6b6e":[1,0,0,1,1,29],
+"classjson_1_1utility_1_1json_object.html#a892221be33efd2082d18b3c5360c7bec":[1,0,0,1,1,31],
+"classjson_1_1utility_1_1json_object.html#a8f45d03f5bede54ca785f21fdc11ae41":[1,0,0,1,1,23],
+"classjson_1_1utility_1_1json_object.html#a91454b0ccbf9e3016bb9d1a3a7f3e409":[1,0,0,1,1,17],
+"classjson_1_1utility_1_1json_object.html#a9427b0184b7dc8ab1527e99422fe44fa":[1,0,0,1,1,19],
 "classjson_1_1utility_1_1json_object.html#a9482b3512c268e6ba157c351f6829b71":[1,0,0,1,1,2],
-"classjson_1_1utility_1_1json_object.html#aa8489ed73d285ba0d172578706e3f460":[1,0,0,1,1,27],
-"classjson_1_1utility_1_1json_object.html#ab1486715ff975a002f9d2041bd35f427":[1,0,0,1,1,6],
-"classjson_1_1utility_1_1json_object.html#ab53545e9f06f1816b2957dfe70b9f11a":[1,0,0,1,1,32],
-"classjson_1_1utility_1_1json_object.html#abfed87b0a6a510467ac50393ce53b0c0":[1,0,0,1,1,35],
-"classjson_1_1utility_1_1json_object.html#ac4c9d994ce1347043481306e1b164bdb":[1,0,0,1,1,13],
+"classjson_1_1utility_1_1json_object.html#a94b6eb7a66bbe05e21dbb9403586f7e9":[1,0,0,1,1,37],
+"classjson_1_1utility_1_1json_object.html#a959fe35ffc1f861f1b84885bab8848a6":[1,0,0,1,1,13],
+"classjson_1_1utility_1_1json_object.html#a9cdd39a00489790321c857b7b5492a10":[1,0,0,1,1,10],
+"classjson_1_1utility_1_1json_object.html#aa8489ed73d285ba0d172578706e3f460":[1,0,0,1,1,26],
+"classjson_1_1utility_1_1json_object.html#aaa17b26aed63c10cb287760559ce1ae6":[1,0,0,1,1,32],
+"classjson_1_1utility_1_1json_object.html#aaaa43ff0df2272147af9173f642d7330":[1,0,0,1,1,40],
+"classjson_1_1utility_1_1json_object.html#ab53545e9f06f1816b2957dfe70b9f11a":[1,0,0,1,1,30],
+"classjson_1_1utility_1_1json_object.html#aba52e50c5572d26e7308ec2e32bcc32b":[1,0,0,1,1,33],
+"classjson_1_1utility_1_1json_object.html#abd2dd20c04a47a19ad0e5e42857a4a1d":[1,0,0,1,1,46],
+"classjson_1_1utility_1_1json_object.html#abf8cc0b2b4cfd44b4838ee7518e160bf":[1,0,0,1,1,25],
+"classjson_1_1utility_1_1json_object.html#ac92e7288687ceefdfe0a25aa496f70d1":[1,0,0,1,1,12],
 "classjson_1_1utility_1_1json_object.html#ad380414c04269e67093770a9ac55f78c":[1,0,0,1,1,1],
-"classjson_1_1utility_1_1json_object.html#ad75e68ccf8d8b9ab1fd425f851348ac5":[1,0,0,1,1,23],
-"classjson_1_1utility_1_1json_object.html#adcfd4efcc44f07a771ae7c59b48f0d4b":[1,0,0,1,1,24],
-"classjson_1_1utility_1_1json_object.html#add836f1ab5bc7397243df440dba65386":[1,0,0,1,1,38],
-"classjson_1_1utility_1_1json_object.html#adfc8fe04ecef682d882aee70ec146bfd":[1,0,0,1,1,16],
+"classjson_1_1utility_1_1json_object.html#ad75e68ccf8d8b9ab1fd425f851348ac5":[1,0,0,1,1,22],
 "classjson_1_1utility_1_1json_object.html#ae60c1e1f744c71dd9311df6fe9cf9873":[1,0,0,1,1,3],
-"classjson_1_1utility_1_1json_object.html#af12e87025aa2935b89ec8bf8b4e0afa7":[1,0,0,1,1,31],
-"classjson_1_1utility_1_1json_object.html#afd52eafb47e65dee4313d6ebb429de77":[1,0,0,1,1,18],
-"classjson_1_1utility_1_1json_object.html#afea5d32506bced68aaaf04c10b25fe10":[1,0,0,1,1,29],
+"classjson_1_1utility_1_1json_object.html#ae6a5869cbaffc6c3d97c2bf70a579fe4":[1,0,0,1,1,39],
+"classjson_1_1utility_1_1json_object.html#afea5d32506bced68aaaf04c10b25fe10":[1,0,0,1,1,28],
 "classjson_1_1utility_1_1json_object_1_1_const_j_s_o_n_iterator.html":[1,0,0,1,1,0],
 "dir_33c78a012c8d08979f57a54a07694c46.html":[2,0,1],
 "dir_68267d1309a1af8e8297ef4c3efbcdba.html":[2,0,0],
