@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['getarray_0',['getarray',['../classjson_1_1_j_s_o_n_parser.html#ab9c08a5659362b930e18487ee2723045',1,'json::JSONParser::getArray()'],['../classjson_1_1utility_1_1json_object.html#ab1486715ff975a002f9d2041bd35f427',1,'json::utility::jsonObject::getArray()']]],
+  ['getarray_0',['getarray',['../classjson_1_1_j_s_o_n_parser.html#a1690f2adb5837e8724af24c11009db66',1,'json::JSONParser::getArray()'],['../classjson_1_1utility_1_1json_object.html#a376b70a297c08788a38b54bd9c64ee8e',1,'json::utility::jsonObject::getArray()']]],
   ['getasboolarray_1',['getAsBoolArray',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#a3845a11097c786bd6ad4a5cb7413d919',1,'json::utility::JSONArrayWrapper']]],
   ['getasdoublearray_2',['getAsDoubleArray',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#a35311f31c943b6a546cf602294b05f90',1,'json::utility::JSONArrayWrapper']]],
   ['getasint64_5ftarray_3',['getAsInt64_tArray',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#a930d80841904b409807e6562430beaef',1,'json::utility::JSONArrayWrapper']]],
@@ -8,17 +8,17 @@ var searchData=
   ['getasobjectarray_5',['getAsObjectArray',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#a001e93cb9ddbcf179c95deb931605b6d',1,'json::utility::JSONArrayWrapper']]],
   ['getasstringarray_6',['getAsStringArray',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#a2a95f7892d6ea85ee5c49135d5b32e43',1,'json::utility::JSONArrayWrapper']]],
   ['getasuint64_5ftarray_7',['getAsUInt64_tArray',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#ae2917d56af64b8e863f272ffbc0d7918',1,'json::utility::JSONArrayWrapper']]],
-  ['getbool_8',['getbool',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#a85e6b36e76afcdb76d29229b89b6a9d4',1,'json::utility::JSONArrayWrapper::getBool()'],['../classjson_1_1_j_s_o_n_parser.html#a0d5793e89a23798caf84ea484dbbaaa7',1,'json::JSONParser::getBool()'],['../classjson_1_1utility_1_1json_object.html#a27b161e7c92fc41b2053281b275b8adb',1,'json::utility::jsonObject::getBool()']]],
-  ['getdouble_9',['getdouble',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#a0336b1d47c3b29f993c7079025c9cd6d',1,'json::utility::JSONArrayWrapper::getDouble()'],['../classjson_1_1_j_s_o_n_parser.html#a09c38da50018f7e9cd5310d05630a05e',1,'json::JSONParser::getDouble()'],['../classjson_1_1utility_1_1json_object.html#a8d55483e4b2113652d4318d6a210910d',1,'json::utility::jsonObject::getDouble()']]],
-  ['getint_10',['getint',['../classjson_1_1_j_s_o_n_parser.html#aebb676dc15e1fa1ba6977a3e4fa18159',1,'json::JSONParser::getInt()'],['../classjson_1_1utility_1_1json_object.html#a5b0954200ef06a70e2cf9b3a4afdf3db',1,'json::utility::jsonObject::getInt()']]],
+  ['getbool_8',['getbool',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#a85e6b36e76afcdb76d29229b89b6a9d4',1,'json::utility::JSONArrayWrapper::getBool()'],['../classjson_1_1_j_s_o_n_parser.html#ab1987fc0f8a9ab7cda6b72af3cde4a1c',1,'json::JSONParser::getBool()'],['../classjson_1_1utility_1_1json_object.html#a61d5124942814d04b15c2559662cb299',1,'json::utility::jsonObject::getBool()']]],
+  ['getdouble_9',['getdouble',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#a0336b1d47c3b29f993c7079025c9cd6d',1,'json::utility::JSONArrayWrapper::getDouble()'],['../classjson_1_1_j_s_o_n_parser.html#a52b34b90e0c81e4885791b792824e503',1,'json::JSONParser::getDouble()'],['../classjson_1_1utility_1_1json_object.html#a199c972a61a209dffdff84d0438671ad',1,'json::utility::jsonObject::getDouble()']]],
+  ['getint_10',['getint',['../classjson_1_1_j_s_o_n_parser.html#a2ebdf48b377a6dbe3f4764a6146c65b2',1,'json::JSONParser::getInt()'],['../classjson_1_1utility_1_1json_object.html#a45b188c694310cc505a979b635d1f6f2',1,'json::utility::jsonObject::getInt()']]],
   ['getint64_5ft_11',['getInt64_t',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#a3fc2731604cfb69acfee3e921c92fd1b',1,'json::utility::JSONArrayWrapper']]],
-  ['getnull_12',['getnull',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#ae4be325fc2faa2b37c60275e7314ae0a',1,'json::utility::JSONArrayWrapper::getNull()'],['../classjson_1_1_j_s_o_n_parser.html#af320cb94a23053295e885e86edd6fa8a',1,'json::JSONParser::getNull()'],['../classjson_1_1utility_1_1json_object.html#a5fa34da0933d80241946a7eb6d361d1c',1,'json::utility::jsonObject::getNull()']]],
-  ['getobject_13',['getobject',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#a14f73a7640e18f5995d62978fcdaabfb',1,'json::utility::JSONArrayWrapper::getObject()'],['../classjson_1_1_j_s_o_n_builder.html#aa3630aeaa94d69b354d962ce8960d3d4',1,'json::JSONBuilder::getObject() const'],['../classjson_1_1_j_s_o_n_builder.html#ad1262ea8b570717e168c50f1ea9734b4',1,'json::JSONBuilder::getObject(utility::jsonObject &amp;object) noexcept'],['../classjson_1_1_j_s_o_n_parser.html#a846ebb23fe7b987cbc05da7af7cd58a0',1,'json::JSONParser::getObject()'],['../classjson_1_1utility_1_1json_object.html#a598e2050cbadc5c695c7737302193b51',1,'json::utility::jsonObject::getObject()']]],
+  ['getnull_12',['getnull',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#ae4be325fc2faa2b37c60275e7314ae0a',1,'json::utility::JSONArrayWrapper::getNull()'],['../classjson_1_1_j_s_o_n_parser.html#aa645009eaee379cbfddb19c31f82fb5a',1,'json::JSONParser::getNull()'],['../classjson_1_1utility_1_1json_object.html#a9cdd39a00489790321c857b7b5492a10',1,'json::utility::jsonObject::getNull()']]],
+  ['getobject_13',['getobject',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#a14f73a7640e18f5995d62978fcdaabfb',1,'json::utility::JSONArrayWrapper::getObject()'],['../classjson_1_1_j_s_o_n_builder.html#aa3630aeaa94d69b354d962ce8960d3d4',1,'json::JSONBuilder::getObject() const'],['../classjson_1_1_j_s_o_n_builder.html#ad1262ea8b570717e168c50f1ea9734b4',1,'json::JSONBuilder::getObject(utility::jsonObject &amp;object) noexcept'],['../classjson_1_1_j_s_o_n_parser.html#a38b1754bcf80b96be4e7953003bffadf',1,'json::JSONParser::getObject()'],['../classjson_1_1utility_1_1json_object.html#a49d5388f6c2b1f7add9df2118ce955ce',1,'json::utility::jsonObject::getObject()']]],
   ['getoffset_14',['getoffset',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#afe23a16435f74e45eb26658b3b330fea',1,'json::utility::JSONArrayWrapper::getOffset() const'],['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#a98792098faf4061192b7ba218c9c1ad3',1,'json::utility::JSONArrayWrapper::getOffset()']]],
   ['getparseddata_15',['getparseddata',['../classjson_1_1_j_s_o_n_parser.html#a4638e39245740b5bd8168f5d8b428728',1,'json::JSONParser::getParsedData() const'],['../classjson_1_1_j_s_o_n_parser.html#a01395b13e66011b067567d97bedea32c',1,'json::JSONParser::getParsedData(utility::jsonObject &amp;object) noexcept']]],
   ['getrawdata_16',['getRawData',['../classjson_1_1_j_s_o_n_parser.html#ae3382fde4943072afef448781533d241',1,'json::JSONParser']]],
-  ['getstring_17',['getstring',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#a701af9690937fec556ca08dd4db8bb9f',1,'json::utility::JSONArrayWrapper::getString()'],['../classjson_1_1_j_s_o_n_parser.html#ac711fd5c6efdfaed2703de7f9b2125e3',1,'json::JSONParser::getString()'],['../classjson_1_1utility_1_1json_object.html#a60c5e1e45ea291be5f3aa67d143d61ce',1,'json::utility::jsonObject::getString()']]],
+  ['getstring_17',['getstring',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#a701af9690937fec556ca08dd4db8bb9f',1,'json::utility::JSONArrayWrapper::getString()'],['../classjson_1_1_j_s_o_n_parser.html#acbbbbcfec6d3b5f1d60c8eff7250b902',1,'json::JSONParser::getString()'],['../classjson_1_1utility_1_1json_object.html#ac92e7288687ceefdfe0a25aa496f70d1',1,'json::utility::jsonObject::getString()']]],
   ['getuint64_5ft_18',['getUInt64_t',['../classjson_1_1utility_1_1_j_s_o_n_array_wrapper.html#a5042e891ab320f418b90f2dfa8caa14e',1,'json::utility::JSONArrayWrapper']]],
-  ['getunsignedint_19',['getunsignedint',['../classjson_1_1_j_s_o_n_parser.html#a6cd09b58620722ad3b18e587e12d7d28',1,'json::JSONParser::getUnsignedInt()'],['../classjson_1_1utility_1_1json_object.html#ac4c9d994ce1347043481306e1b164bdb',1,'json::utility::jsonObject::getUnsignedInt()']]],
-  ['getvalue_20',['getValue',['../classjson_1_1_j_s_o_n_parser.html#a96fbc3ab20618e40747e7c9ca41bf0f6',1,'json::JSONParser']]]
+  ['getunsignedint_19',['getunsignedint',['../classjson_1_1_j_s_o_n_parser.html#a3e7447a67746f778a3587365c40e7332',1,'json::JSONParser::getUnsignedInt()'],['../classjson_1_1utility_1_1json_object.html#a959fe35ffc1f861f1b84885bab8848a6',1,'json::utility::jsonObject::getUnsignedInt()']]],
+  ['getvalue_20',['getValue',['../classjson_1_1_j_s_o_n_parser.html#a2a51366349730b9e7ea2da82603904d9',1,'json::JSONParser']]]
 ];
