@@ -1,13 +1,15 @@
 #include "CantFindValueException.h"
 
+#include <format>
+
 using namespace std;
 
 namespace json
 {
 	namespace exceptions
 	{
-		CantFindValueException::CantFindValueException(const string& key) :
-			BaseJSONException("Can't find JSON value with key \"" + key + '"')
+		CantFindValueException::CantFindValueException(string_view key) :
+			BaseJSONException(format(R"(Can't find JSON value with key "{}")", key))
 		{
 
 		}
